// apps/feeds/frontend/src/app/graphql/types.ts
export type Maybe<T> = T | null
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string
  String: string
  Boolean: boolean
  Int: number
  Float: number
  LocalDateTime: any
}

export type Article = {
  __typename?: 'Article'
  id: Scalars['ID']
  link: Scalars['String']
  title: Scalars['String']
  author: Scalars['String']
  content: Scalars['String']
  publishedDate: Scalars['LocalDateTime']
  feed: Feed
  unread: Scalars['Boolean']
}

export type ArticleFilterCriteria = {
  feedIds?: Maybe<Array<Scalars['ID']>>
  offset?: Maybe<Scalars['Int']>
  limit?: Maybe<Scalars['Int']>
}

export type Feed = {
  __typename?: 'Feed'
  id: Scalars['ID']
  uri: Scalars['String']
  title: Scalars['String']
  icon?: Maybe<Scalars['String']>
  articles: Array<Article>
  group?: Maybe<Group>
  unreadCount?: Maybe<Scalars['Int']>
  totalCount?: Maybe<Scalars['Int']>
  hasFetchError: Scalars['Boolean']
}

export type FeedArticlesArgs = {
  unread?: Maybe<Scalars['Boolean']>
  offset?: Maybe<Scalars['Int']>
  limit?: Maybe<Scalars['Int']>
}

export type FeedInput = {
  uri: Scalars['String']
  title?: Maybe<Scalars['String']>
}

export type Group = {
  __typename?: 'Group'
  id: Scalars['ID']
  name: Scalars['String']
  feeds: Array<Feed>
}
// ...
export type Query = {
  __typename?: 'Query'
  articles: Array<Article>
  feeds: Array<Feed>
  groups: Array<Group>
  authorizationUri: Scalars['String']
  login?: Maybe<Token>
  me?: Maybe<User>
}

export type QueryArticlesArgs = {
  filter: ArticleFilterCriteria
}

export type QueryLoginArgs = {
  code: Scalars['String']
}

export type Subscription = {
  __typename?: 'Subscription'
  feed?: Maybe<Feed>
}

export type Token = {
  __typename?: 'Token'
  access_token: Scalars['String']
  refresh_token: Scalars['String']
  scope: Scalars['String']
  token_type: Scalars['String']
}

export type User = {
  __typename?: 'User'
  username: Scalars['String']
  name?: Maybe<Scalars['String']>
  avatar_url?: Maybe<Scalars['String']>
  email?: Maybe<Scalars['String']>
}

export type ArticleFieldsFragment = { __typename: 'Article' } & Pick<
  Article,
  'id' | 'title' | 'link' | 'author' | 'content' | 'publishedDate' | 'unread'
> & { feed: { __typename: 'Feed' } & Pick<Feed, 'id'> }

export type FeedFieldsFragment = { __typename: 'Feed' } & Pick<
  Feed,
  | 'id'
  | 'uri'
  | 'title'
  | 'icon'
  | 'unreadCount'
  | 'totalCount'
  | 'hasFetchError'
> & { group: Maybe<{ __typename: 'Group' } & Pick<Group, 'id' | 'name'>> }

export type GroupFieldsFragment = { __typename: 'Group' } & Pick<
  Group,
  'id' | 'name'
> & { feeds: Array<{ __typename: 'Feed' } & Pick<Feed, 'id'>> }

export type TokenFieldsFragment = { __typename?: 'Token' } & Pick<
  Token,
  'access_token' | 'refresh_token' | 'scope' | 'token_type'
>

export type UserFieldsFragment = { __typename?: 'User' } & Pick<
  User,
  'username' | 'name' | 'avatar_url' | 'email'
>

export type FeedsAndGroupsQueryVariables = {}

export type FeedsAndGroupsQuery = { __typename?: 'Query' } & {
  feeds: Array<{ __typename?: 'Feed' } & FeedFieldsFragment>
  groups: Array<{ __typename?: 'Group' } & GroupFieldsFragment>
}